Move mount tests to integration directory
