+++ apps/test/unit/code-studio/pd/application/TeacherApplicationTest.js
@@ -1,76 +1,32 @@
 import React from 'react';
 import {expect} from 'chai';
-import {mount, shallow} from 'enzyme';
+import {shallow} from 'enzyme';
 import sinon from 'sinon';
 import {isolateComponent} from 'isolate-react';
 import firehoseClient from '@cdo/apps/lib/util/firehose';
-import {PageLabels} from '@cdo/apps/generated/pd/teacherApplicationConstants';
 import TeacherApplication from '@cdo/apps/code-studio/pd/application/teacher/TeacherApplication';
 
 describe('TeacherApplication', () => {
-  const fakeOptionKeys = Object.values(PageLabels).reduce(
-    (acc, cur) => acc.concat(Object.keys(cur)),
-    []
-  );
-  const fakeOptions = fakeOptionKeys.reduce(
-    (acc, cur) => ({...acc, [cur]: ['1', '2', '3']}),
-    {}
-  );
   const defaultProps = {
     apiEndpoint: '/path/to/endpoint',
-    requiredFields: [],
-    options: fakeOptions,
+    options: {},
     accountEmail: 'user@email.com',
     userId: 1
   };
 
-  let ga;
+  let fc;
 
   beforeEach(() => {
-    ga = sinon.fake();
-    window.ga = ga;
-    window.sessionStorage.removeItem('TeacherApplication');
+    fc = sinon.stub(firehoseClient, 'putRecord');
   });
 
   afterEach(() => {
-    window.ga = undefined;
-    sinon.restore();
-    window.sessionStorage.removeItem('TeacherApplication');
+    fc.restore();
   });
 
-  it('Does not set schoolId if not provided', () => {
-    const page = mount(<TeacherApplication {...defaultProps} />);
-    expect(page.find('SchoolAutocompleteDropdown').prop('value')).to.equal(
-      undefined
-    );
-  });
-  it('Sets the school dropdown value from props', () => {
-    const page = mount(<TeacherApplication {...defaultProps} schoolId="50" />);
-    expect(page.find('SchoolAutocompleteDropdown').prop('value')).to.equal(
-      '50'
-    );
-  });
-  it('Sets the school dropdown value from storage', () => {
-    sinon
-      .stub(window.sessionStorage, 'getItem')
-      .withArgs('TeacherApplication')
-      .returns(JSON.stringify({data: {school: '25'}}));
-    const page = mount(<TeacherApplication {...defaultProps} />);
-    expect(page.find('SchoolAutocompleteDropdown').prop('value')).to.equal(
-      '25'
-    );
-    window.sessionStorage.getItem.restore();
-  });
   it('Logs user id on initialization', () => {
-    const fc = sinon.stub(firehoseClient, 'putRecord');
     shallow(<TeacherApplication {...defaultProps} />);
     expect(fc.calledWith(sinon.match({userId: 1})));
-  });
-  it('Reports to google analytics', () => {
-    const form = mount(<TeacherApplication {...defaultProps} />);
-    const nextButton = form.find('#next').first();
-    nextButton.simulate('click');
-    expect(ga.getCall(2).calledWith());
   });
 
   describe('getInitialData', () => {
